pyproject.toml
[project]
name = "crawler"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "playwright[Chromium]>=1.49.0",
    "python-decouple>=3.8",
]
